Cambio tipo de dato de int a long para buscar paciente por dni y quito prueba de excepcion de diagnostico ya agregado

+++ src/main/java/com/example/is_tfi/repositorio/RepositorioPaciente.java
@@ -6,5 +6,5 @@ import java.util.Optional;
 
 public interface RepositorioPaciente {
     void guardarPaciente(Paciente paciente);
-    Optional<Paciente> buscarPacientePorDni(int dni);
+    Optional<Paciente> buscarPacientePorDni(Long dni);
 }
